refactor: 패키지 구조 나누기
