+++ src/main/kotlin/domain/lottogenerator/ManualLottoGenerator.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.lottogenerator
 
 import domain.model.lotto.Lotto
 import domain.model.lotto.LottoNumber
