+++ src/test/kotlin/lottogenerator/AutomaticLottoGeneratorTest.kt
@@ -1,6 +1,6 @@
 package lottogenerator
 
-import domain.AutomaticLottoGenerator
+import domain.lottogenerator.AutomaticLottoGenerator
 import domain.model.lotto.Lotto
 import org.assertj.core.api.Java6Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
