+++ src/test/kotlin/lottogenerator/LottoSellerTest.kt
@@ -5,10 +5,11 @@ import domain.model.PurchaseMoney
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
+import org.junit.jupiter.api.assertThrows
 
 class LottoSellerTest {
 
-    lateinit var lottoSeller: LottoSeller
+    private lateinit var lottoSeller: LottoSeller
 
     @BeforeEach
     fun setUp() {
@@ -24,13 +25,13 @@ class LottoSellerTest {
 
     @Test
     fun `사용자가 입력한 돈이 1000원 단위이 아닌 경우 예외가 발생`(){
-        assertDoesNotThrow {
-            lottoSeller.receiveMoneyFromCustomer(PurchaseMoney(10000), 3)
+        assertThrows<IllegalArgumentException> {
+            lottoSeller.receiveMoneyFromCustomer(PurchaseMoney(10500), 3)
         }
     }
 
     @Test
-    fun `로또 6개중 2개를`() {
+    fun `로또 6개중 3개를 `() {
 
     }
 }
